Refactoring amount from double to AtomicLong (now its in cents). Create test class for multithreading (my first time?)

+++ src/main/java/com/github/zeykrus/bankpet/interfaces/CreditAllowed.java
@@ -2,8 +2,8 @@ package com.github.zeykrus.bankpet.interfaces;
 
 public interface CreditAllowed extends PeriodicOperation {
     double DEFAULT_CREDIT_PERCENT = 0.05;
-    double DEFAULT_CREDIT_LIMIT = 1000;
+    long DEFAULT_CREDIT_LIMIT = 1000;
 
-    void setCreditLimit(double credit);
-    double getCreditLimit();
+    void setCreditLimit(long credit);
+    long getCreditLimit();
 }
