+++ src/test/java/com/github/zeykrus/bankpet/account/SavingsAccountTest.java
@@ -2,10 +2,8 @@ package com.github.zeykrus.bankpet.account;
 
 import com.github.zeykrus.bankpet.FinanceCoreEngine;
 import com.github.zeykrus.bankpet.TestConstants;
-import com.github.zeykrus.bankpet.exception.InsufficientFundsException;
-import com.github.zeykrus.bankpet.model.Bank;
+import com.github.zeykrus.bankpet.services.Bank;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class SavingsAccountTest extends AccountTest<SavingsAccount> {
@@ -18,28 +16,28 @@ public class SavingsAccountTest extends AccountTest<SavingsAccount> {
 
     @Test
     void canWithdrawPositive() {
-        double amount = account.getBalance() / 2; //меньше, чем есть
+        long amount = account.getBalance() / 2; //меньше, чем есть
 
         Assertions.assertTrue(account.canWithdraw(amount),"Ожидался положительный исход возможности снятия средств");
     }
 
     @Test
     void canWithdrawNegative() {
-        double amount = account.getBalance() * 2; //больше, чем есть
+        long amount = account.getBalance() * 2; //больше, чем есть
 
         Assertions.assertFalse(account.canWithdraw(amount),"Ожидался отрицательный исход возможности снятия средств");
     }
 
     @Test
     void notEnoughWhenEnough() {
-        double amount = account.getBalance() / 2; //меньше, чем есть
+        long amount = account.getBalance() / 2; //меньше, чем есть
 
         Assertions.assertEquals(0, account.notEnough(amount), "Ожидалось, что денег достаточно (не хватает 0)");
     }
 
     @Test
     void notEnoughPositive() {
-        double amount = account.getBalance() + TestConstants.POSITIVE_AMOUNT;
+        long amount = account.getBalance() + TestConstants.POSITIVE_AMOUNT;
 
         Assertions.assertEquals(TestConstants.POSITIVE_AMOUNT, account.notEnough(amount), "Ожидалось нехватка конкретной суммы средств");
     }
@@ -47,7 +45,7 @@ public class SavingsAccountTest extends AccountTest<SavingsAccount> {
     @Test
     void constructorTest() {
         String owner = "Owner1";
-        double initBalance = TestConstants.POSITIVE_AMOUNT;
+        long initBalance = TestConstants.POSITIVE_AMOUNT;
         int number = 10;
 
         SavingsAccount acc = new SavingsAccount(bank, number, owner, initBalance);
